requirements.txt
requests==2.31.0
urllib3==1.26.20
psutil==5.9.8
openai==1.3.0
python-dotenv==1.0.0